Fix eloop delays by using a dummy socket
Eloop framework in wpa_supplicant uses select with minimal timeout of
all registered users, but by default it is 10secs (periodic cleanup
task), so, in case of no other events all registered users will be
delayed by 10secs.

In Linux, select exits without waiting for full 10secs as for there
is a handler registered for NL80211 sockets and whenever there is a
event from Kernel select exits and processes expired events immediately.

In Zephyr, we don't have such mechanism as we use direction function
calls between kernel and wpa_supplicant, so, add a dummy socket and
register it with Eloop and post a dummy a message on any interesting
events like shell command, event from the driver etc. to exit select for
faster events processing, this signficantly improves the association
time from 30s to 5s.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>