[nrf noup] mbedtls: Fix SAE-H2E sliding window exponentiation
For some passwords using SAE-H2E the Mbedtls PWE derivation fails with
invalid public key (-0x4C80) i.e., the point is not on the elliptic
curve.

This is because of the MbedTLS API mbedtls_mpi_exp_mod (called from
sae_derive_pwe_from_pt_ecc using crypto_ec_point_mul) warns about using
aliasing b/w its input and output arguments, but the general style of
WPA supplicant is to pass same pointer as input and output, this causes
wrong computation, but this is only checked when deriving PWE not when
computing the point.

The dependency of password comes in with pwd_seed which is based on both
SSID and password (see "sae_derive_pt_ecc" for details) as these are
used to calculate hash which is used to map to an element on the curve.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>
(cherry picked from commit d59f86315916ac09267b8bc871c1ae31b507ec4c)