[nrf noup] zephyr: Fix handling of IEs in authentication
When copying event from driver and postnig to WPA supplicant, the IE's
are not copied, so, if WPA supplicant takes longer to process the event
the driver frees the IEs causing invalid data to be processed by WPA
supplicant. This is a common issue to all events.

For now Authentication events are being affected due to WPA3-SAE taking
longer to process, so, fix only Authentication events for now as a
generic fix needs more work and thought.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>
(cherry picked from commit e0c536c1535dec7948886efe2702a1a312702251)