Use unbuffered logging for debugs
WPA supplicant debugs are excessive and the incoming rate is high esp.
during connection, so, even though we increase the log buffer size, we
still miss quite a few events.

So, disable buffered logging for debugs, one caveat is that we lose the
timestamp that comes with LOG_ macros and also the prints might be
jumbled but this is still preferred over losing events.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>