Fix shell hang when memory is low
When available HEAP is low, the allocation failure to receive the
message from control interface can cause the control interface socket to
be always "ready" causing the WPA supplicant eloop to a busy loop and no
other threads are run (The thread that implements control interface
command timeout). If this is from shell, then it results in a shell hang
forever.

Use stack to avoid this busy loop.

Fixes WSP-137.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>