[nrf fromlist] zephyr: De-couple MBO and WNM
MBO is mandatory for WFA certification but WNM is not, so, in order to
have a lean build for WFA certification only, de-couple WNM from MBO so
that it can be compiled out.

WNM is still auto-selected automatically when MBO is enabled, this
behaviour is unchanged.

Upstream PR: http://lists.infradead.org/pipermail/hostap/2023-November/042013.html

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>