[nrf noup] zephyr: Add length check for response
fixup! [nrf noup] zephyr: Add support for WPA CLI zephyr

As length is output of number of bytes received from socket we should
handle zero bytes as well, since we use stream sockets this happens when
the other end is shutdown.

Without this fix in a resource constrained envrionment len can become
zero and as its unsigned it overflows causing a crash.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>