[nrf noup] zephyr: Fix argument count for wpa_cli
fixup! [nrf noup] zephyr: Add support for WPA CLI zephyr

For wpa_cli command the argument count was off by 1 as we remove the
command itself and pass rest of the arguments to the handler. This still
works fine for most, but for "interface_add" it relies of argument count
for preparing command over control interface and ends up setting a L2
bridge interface which is invalid and hence causing issues.

Fixes SHEL-2196.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>