[nrf noup] zephyr: Fix MBO and GAS dependency
fixup! [nrf noup] zephyr: Add support for Zephyr RTOS

commit eb6119524 ("[nrf fromlist] zephyr: Fix MBO build with GAS
dependency") fixes the dependency in opensource acceptable way, but
enable of GAS pulls in other modules costing memory.

MBO uses GAS query only for HS2.0 which isn't enabled in Zephyr, so,
instead of implicitly enabling GAS in Zephyr for MBO, we protect this
code with GAS, which avoids pulling in code and saves memory.

This was working till date because Zephyr toolchain silently removed the
uncalled functions.

In the future if we add support for HS2.0 then GAS will be automatically
enabled and this still works.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>