[nrf noup] zephyr: Fix message processing
fixup! [nrf noup] Monitor supplicant state and inform applications

Incoming message has a pre-defined structure and isn't a arbitrary blob,
so, use the structure to parse properly.

This was working earlier as the payload was in the beginning of the
structure, but if we ever end up with a longer message then we would see
issues.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>