[nrf noup] zephyr: Fix reason code for locally generated disconnection
fixup! [nrf noup] zephyr: Implement connect and disconnect events to Zephyr

In case a use issues a disconnect for some reason, it will still send
IEEE 802.11 reason code 3, which will show up as a Failure (non-zero
status) in Zephyr, this should only be applicable if AP sends
de-authentication with reason code as "3".

Fix locally generated case to return success (the enum should also be
renamed from "unspecified" to "local", but that will be handled in
another PR).

Fixes SHEL-2561.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>