[nrf noup] zephyr: Disable WNM pre-scan check
Pre-scan check modifies the behaviour when no candidate is found, with
this check enabled a fresh and full scan is forced which triggers the
disassociation if the timeout is closer to scan time.

Disable this check as a workaround to establish pre-upmerge WNM
behaviour in case no candidate is found (WFA QT uses single AP, so,
you will never find a candidate).

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>